[IMP] estate,estate_account: created a compute method, created a bridge module.
Created one compute method, in which when garden field is set to true, the value
 of garden area will be 10 and garden orientation will be 'north'.
created a ondelete method, in which we will not be able to delete a property
with state  new and offer received.
created a 'create method' in which the offer can not be created with less price
than the existing one.
Created a bridge module estate_account.