[IMP] real_estate: when property sold, invoice will be generated.
Extended the functionality of action_sold_btn in estate_property Module. when s
alesperson accept the offer and sold the property to the buyer, Invoice will be
automatic generated in Invoicing module. Also created view and action for sold p
roperties, which display all info about buyers when property is sold.