[ADD] real_estate : Linked Two modules estate_account and estate_project
Linked two modules estate account and estate project with our
real estate module.

Estate Account is linked to create invoice for the sold property.

Estate Project is linked to create a project named estate property
and a task for maintenance

Did some changes which krina mam suggested in the draft