[IMP] estate_account: invoice creation
Create a invoice successfully.
Done a Exercise of  6% of the selling price. an additional 100.00 from
administrative fees.get the idea about the Command.create method.
Create a estate_project module.
inherit with estate property.
Implement the bridge module with project.in that create a project and for that
project create a multiple task when property is sold.