[ADD] estate_project: create bridge module between estate and project
In the estate_project bridge module, we created a new project and its task
through estate property. In the estate property module when one property is sold
then corresponding projects and tasks are created. The buyer and Seller of the
property will be shown in the project details.Many2many tags are also shown in
the project task.