[ADD] estate : added-required-constraints-order-of-properties-and-status-bar
Added sql and python constraints which rejects invalid inputs and lets seller
accept offers only if they are greater than 90% of the expected price.  Added
ordering which shows properties in ascending order of their price and offers
from highest to lowest. Added status bar which shows the state of the property
and added sequencing of property types