[IMP] real_estate: acl and record rules
implement the Access Rights and record rules in real estate module.
create agent and manager user to know how can we actually provides security to
models or fields or how we differentiate users.
get to know about the sudo() method to  bypass access rights and rules when
creating the invoice.