[IMP] estate:  Improved UI, Logic for the fields,Sequence for custom order.
Purpose: Color were added in the User Interface to make it more intuitive ,
Sequence was added for the easy of use for the user,Custom filter was added
for better search,some fields attribute are connected with logic for easy to use.

Work Done: Color were added in tags, offers,property tree view. sequence and
handle is added with default order to give the user a better experience or
custom sorting. A custom filter was added in living area for a better searching.
Char field which are linked with some field are now only visible when certain
condition are met ex : garden and once a property reach the sold stage
the sold and cancel button are no more visible. related fields are added
but the stat button work is in progress.