[ADD] real_estate_module:  record_rules, security_override
Updated access rights for the agents
Added record rules for the records for the agents as well as for the manager.
We have to define the record rule for the model ir.rule and defined domain force
so that only users with their respective names assigned to the property will be
visible to them(agents).
Also added security override for the invoicing feature for the agents using
sudo function.