[ADD] real_estate_module:  programmatically checking security
Added check.access.rights() method and check.access.rules() method.
The CRUD name is passed as the argument to these methods.

Through these methods it checks whether the user accessing the record
is allowed to do the specified CRUD operation(which is passed as an argument)
or not.

If the user is not allowed with the specified CRUD operation then it will
raise a access error.