[ADD] real_estate:  added wizard through which offers can be added
Added wizards for the real_estate module.
When agent wants to add offers to all the available properties then he can do
with the help of wizard.
We have to create a wizard model which inherits from transient model - where
data is non-permanent and then we have to define its action and view.
Wizard form view is attached to the button of the estate.property tree view.
When agent selects single or multiple properties, the add offer button gets
activated and then the agent can upload offers to all the selected properties.

self.env is used to read the properties inside estate.property and also used
active_ids which will show only the properties which are selected.
To add the offer, create method is used along with self.env for estate offers