[ADD] estate:  Added access controls
Added groups for agent and manager and added access rights in which Employees
who are not at least real-estate agents will not see the real-estate application.
and Real-estate agents will not be able to update the property types or tags.
Also added record rules in which agents will not be able to see the properties
exclusive to their colleagues; but managers will still be able to see everything.