[IMP] estate : Added bridge modules and views
Additions:
Bridge modules to connect the self practice with project and accounts.
'estate_project' to bridge with project module. 'estate_account' to bridge with
accounts module.

Purpose:
The account bridge would enable us to create an invoice, automatically, when a
property is sold.
The project would enable us to create a new project and a new task for when
a property is sold.